requirements.txt
pyautogui==0.9.53

wikipedia==1.4.0
wolframalpha==5.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pyttsx3==2.90
psutil==5.9.5
screen-brightness-control==0.8.3
opencv-python==4.8.1.78
plyer==2.1.0
speedtest-cli==2.1.3
pywin32==306
winshell==0.6